+++ internal/metrics/metrics.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"math"
 	"sort"
+	"strconv"
 
 	"github.com/getsentry/sentry-go"
 	"github.com/getsentry/vroom/internal/chunk"
@@ -111,17 +112,40 @@ func quantile(values []uint64, q float64) (uint64, error) {
 	return values[index], nil
 }
 
+func ExtractFunctionsFromCallTreesForThread(
+	callTreesForThread []*nodetree.Node,
+) []nodetree.CallTreeFunction {
+	functions := make(map[uint32]nodetree.CallTreeFunction, 0)
+
+	for _, callTree := range callTreesForThread {
+		callTree.CollectFunctions(functions, "")
+	}
+
+	return mergeAndSortFunctions(functions)
+}
+
 func ExtractFunctionsFromCallTrees[T comparable](
 	callTrees map[T][]*nodetree.Node,
 ) []nodetree.CallTreeFunction {
 	functions := make(map[uint32]nodetree.CallTreeFunction, 0)
-
-	for _, callTreesForThread := range callTrees {
+	for tid, callTreesForThread := range callTrees {
+		threadID := ""
+		if t, ok := any(tid).(string); ok {
+			threadID = t
+		} else if t, ok := any(tid).(uint64); ok {
+			threadID = strconv.FormatUint(t, 10)
+		}
 		for _, callTree := range callTreesForThread {
-			callTree.CollectFunctions(functions)
+			callTree.CollectFunctions(functions, threadID)
 		}
 	}
 
+	return mergeAndSortFunctions(functions)
+}
+
+func mergeAndSortFunctions(
+	functions map[uint32]nodetree.CallTreeFunction,
+) []nodetree.CallTreeFunction {
 	functionsList := make([]nodetree.CallTreeFunction, 0, len(functions))
 	for _, function := range functions {
 		if function.SampleCount <= 1 {